arm64: dts: qcom: sc7180-trogdor: Connect DP to cros-ec-typec
Qualcomm's SC7180 SoC has one DP PHY that can output 4 lanes of DP
through the QMP "combo" PHY that combines both USB3 and DP PHYs. The QMP
PHY configuration is fixed on Trogdor so that of the 4 total lanes 2
lanes of DP are dedicated for either of the two usb-c-connectors on the
board and the other 2 lanes are dedicated to USB3 which goes to a USB3
hub. Connect the DP lanes to the EC node to complete the graph.

Cc: <cros-qcom-dts-watchers@chromium.org>
Cc: Bjorn Andersson <andersson@kernel.org>
Cc: Konrad Dybcio <konradybcio@kernel.org>
Cc: Rob Herring <robh+dt@kernel.org>
Cc: Krzysztof Kozlowski <krzysztof.kozlowski+dt@linaro.org>
Cc: Conor Dooley <conor+dt@kernel.org>
Cc: <linux-arm-msm@vger.kernel.org>
Cc: <devicetree@vger.kernel.org>
Cc: Pin-yen Lin <treapking@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>